Change arg type

+++ provision/include/virgil/iot/provision/provision.h
@@ -123,7 +123,7 @@ typedef struct __attribute__((__packed__)) {
  * @return              return true if slot has found for element_id
  */
 bool
-vs_provision_get_slot_num(vs_provision_element_id_e id, size_t *slot);
+vs_provision_get_slot_num(vs_provision_element_id_e id, uint16_t *slot);
 
 bool
 vs_provision_search_hl_pubkey(vs_key_type_e key_type, vs_hsm_keypair_type_e ec_type, uint8_t *key, uint16_t key_sz);
